pyproject.toml
[project]
name = "ua-news-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ua_news_bot"]

[tool.hatch.build.targets.sdist]
include = ["src/ua_news_bot", "README.md"]
